The HID Reporting system is moved to using RingBuffer.
There was a pesky bug in ring_buffer, the starting indices for the
sections of the ring were not modded over when the ring_size was
trimmed. Meaning, if the state was (Empty(10,8),Queued(10,0),TXing(2,8))
It would be shrunk to (Empty(0,2),Queued(10,0),TXing(2,8)) instead of
(Empty(0,2),Queued(0,0),TXing(2,8)). Both represent the same state
but the starting index of Queued should be 0 to satisfy boundary
conditions in other functions.